Update csce322h0mework01part01.g4
I needed the Lexer to stop scanning when an Abomination is met. So as opposed to the previous commit where my parser was making the decision, I am using my LEXER to stop the program if an invalid token is found. But to print it out, I had to assign everything other than the valid input to be a token. 
It isn't strictly speaking "correct" but it does the trick.